chore(build): Add support for drift detection of generated code (#22)
* chore(build): Add support for drift detection of generated code

This adds a new workflow that checks that generated code remains the same when go generate is run again. It currently needs to be specifically enabled by adding //check-for-changes to the top of the generator .hcl file.

It also adds drift detection to all existing files and fixes any drift that had already occurred. To get the script to work also involved bringing the directory structure in line with other providers.